Add random text generator and fix Linux support

+++ include/XML.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "FileHolder.h"
+#include "Types.h"
 
 
 class XMLEntry {
@@ -8,7 +9,7 @@ public:
     XMLEntry();
     XMLEntry(FileHolder* pFileHolder);
 
-    bool findEntry(const char* pEntryName, unsigned long pOffset, bool pOnlySearchForEntryStart = false);
+    bool findEntry(const char* pEntryName, u32 pOffset, bool pOnlySearchForEntryStart = false);
     char* getEndOfLongEntry(char* pXmlParser);
     bool getEntryInfo(const char* pEntryInfoName, char* pOutputString);
     bool getNextEntry(const char* pEntryName);
@@ -29,17 +30,17 @@ public:
     XMLSubEntry(const XMLEntry* pXmlParentEntry);
 
     bool findSubEntry(const char* pEntryName);
-    bool findSubEntry(const char* pEntryName, unsigned long pOffset);
+    bool findSubEntry(const char* pEntryName, u32 pOffset);
 
     const XMLEntry* mXmlParentEntry;
 };
 
 namespace XmlUtil {
-    unsigned long getNumberOfEntries(const char* pEntryName, FileHolder* pXmlFile);
-    bool getEntryByIndex(const char* pEntryName, unsigned long pIndex, FileHolder* pXmlFile, XMLEntry* pXmlEntry);
-    unsigned long countLines(const char* pStartString, const char* pEndString);
+    u32 getNumberOfEntries(const char* pEntryName, FileHolder* pXmlFile);
+    bool getEntryByIndex(const char* pEntryName, u32 pIndex, FileHolder* pXmlFile, XMLEntry* pXmlEntry);
+    u32 countLines(const char* pStartString, const char* pEndString);
     bool isValidStringEndChar(char);
     void extractStringFromQuotes(const char* pInput, char* pOutput);
     bool isEqualEntryName(const char* pEntryName, const char* pXmlEntry);
-    long getEntryIndexByEntryNameAndInfo(const char* pEntryName, const char* pInfoName, const char* pInfoValue, FileHolder* pXmlFile);
+    s32 getEntryIndexByEntryNameAndInfo(const char* pEntryName, const char* pInfoName, const char* pInfoValue, FileHolder* pXmlFile);
 }
